Added features
The app can now remember the state (on/off, hue saturation and level) of the lights individually before the display is triggered and revert the lights to their previous settings after the display is complete.

Weather data is no longer retrieved at the time of the motion event. It is now fetched every 15 minutes and cached locally to improve response time.

Added an option to display a button over the icon to trigger the weather display on demand. The button was there already, there's now an option to not show it.